Move all logic so far to Ecosystem::Archive::Update
Turns out that we probably want to run any update logic in a separate
process anyway because of memory and CPU requirements.  So for all of
the introspection, we would probably just run off of the META info as
published in the REA.  Which will be done by the Ecosystem::Archive
module (which is an empty shell now, but at least updating still appears
to work).